gmail_auto_label_by_sender/requirements.txt
psycopg2-binary==2.9.9
openpyxl==3.1.2


